Add files via upload
printjob.py -- the source, of course.
analysis.lyx -- LyX source with the analysis of the problem and math underlying the code. It includes a listing (TODO: figure out how to make a float include the source, so it needn't change just because the Python source code changes.)
analysis.pdf -- strictly speaking a derived object and thus something that shouldn't be under version control, but we put it here for those who don't have LyX. (That said, if you don't have LyX, you should get it. https://www.lyx.org/ is the place to go.)